+++ web/documentserver-example/php/src/configuration/ConfigurationManager.php
@@ -73,7 +73,7 @@ class ConfigurationManager
     {
         $serverURL = $this->documentServerPrivateURL();
         $path = getenv('DOCUMENT_SERVER_COMMAND_PATH')
-            ?: 'coauthoring/CommandService.ashx';
+            ?: 'command';
         return $serverURL->joinPath($path);
     }